feat: use DELETE and PATCH

+++ src/Controller/ProjectController.php
@@ -50,28 +50,32 @@ final class ProjectController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}/edit', name: 'app_project_edit', methods: ['GET', 'POST'])]
+    #[Route('/{id}/edit', name: 'app_project_edit', methods: ['GET', 'POST', 'PATCH'])]
     public function edit(Request $request, Project $project, EntityManagerInterface $entityManager): Response
     {
         $form = $this->createForm(ProjectType::class, $project);
         $form->handleRequest($request);
-
+    
         if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager->flush();
-
-            return $this->redirectToRoute('app_project_index', [], Response::HTTP_SEE_OTHER);
+            dump($request->getMethod());
+            if ($request->getMethod() === 'PATCH') {
+                $entityManager->flush();
+                return $this->redirectToRoute('app_project_index', [], Response::HTTP_SEE_OTHER);
+            }
         }
-
+    
         return $this->render('project/edit.html.twig', [
             'project' => $project,
             'form' => $form,
         ]);
     }
+    
+    
 
-    #[Route('/{id}', name: 'app_project_delete', methods: ['POST'])]
+    #[Route('/{id}', name: 'app_project_delete', methods: ['POST', 'DELETE'])]
     public function delete(Request $request, Project $project, EntityManagerInterface $entityManager): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$project->getId(), $request->getPayload()->getString('_token'))) {
+        if ($request->getMethod() == 'DELETE' && $this->isCsrfTokenValid('delete'.$project->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($project);
             $entityManager->flush();
         }
